Refactored some season methods into the seadons calculator module
